backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.118.2
graphql-core==3.2.6
h11==0.16.0
idna==3.10
lia-web==0.2.3
packaging==25.0
passlib==1.7.4
pydantic==2.12.0
pydantic_core==2.41.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.27
starlette==0.48.0
strawberry-graphql==0.283.2
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
